Refactored Advanced page to have each tab as separate component
